# packages/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SheepHub API",
    "version": "1.0.0",
    "description": "Hono-based API server for SheepHub providing translation and quality checks via Vertex AI."
  },
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "components": {
    "schemas": {
      "GreetResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "description": "Success or error status"
          },
          "model_info": {
            "type": [
              "string",
              "null"
            ],
            "description": "Information about the model"
          },
          "error": {
            "type": [
              "string",
              "null"
            ],
            "description": "Error message if any"
          }
        },
        "required": [
          "status"
        ]
      },
      "ResultResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "description": "Success or error status"
          },
          "result": {
            "type": [
              "string",
              "null"
            ],
            "description": "Processing result text"
          },
          "error": {
            "type": [
              "string",
              "null"
            ],
            "description": "Error message if any"
          }
        },
        "required": [
          "status"
        ]
      },
      "InitPromptRequest": {
        "type": "object",
        "properties": {
          "system_instruction": {
            "type": "string",
            "description": "System instruction prompt for Gemini"
          },
          "display_name": {
            "type": "string",
            "description": "Display name for cached content"
          }
        },
        "required": [
          "system_instruction",
          "display_name"
        ]
      },
      "DeleteCacheRequest": {
        "type": "object",
        "properties": {
          "cache_name": {
            "type": [
              "string",
              "null"
            ],
            "description": "Name of the cache to delete"
          }
        }
      },
      "TaskResponse": {
        "type": "object",
        "properties": {
          "task_id": {
            "type": "string",
            "description": "The UUID of the spawned background task"
          },
          "status": {
            "type": "string",
            "description": "Status of the task"
          }
        },
        "required": [
          "task_id",
          "status"
        ]
      },
      "UserRequest": {
        "type": "object",
        "properties": {
          "chunk": {
            "type": "string",
            "description": "The input JSONL data to process"
          },
          "prompt": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional user override prompt"
          },
          "cache_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional cache name/ID to use"
          }
        },
        "required": [
          "chunk"
        ]
      },
      "RequestBody": {
        "type": "object",
        "properties": {
          "chunk": {
            "type": "string",
            "description": "The input JSONL data to process"
          }
        },
        "required": [
          "chunk"
        ]
      }
    },
    "parameters": {}
  },
  "paths": {
    "/verify_connection": {
      "get": {
        "summary": "Verify Connection",
        "description": "Endpoint to verify that the API server is running and accessible.",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "status",
                    "message"
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Authentication failure",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/tasks/{task_id}": {
      "get": {
        "summary": "Get Task Result",
        "description": "Retrieve the status and result of a background task.",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "The UUID of the task"
            },
            "required": true,
            "name": "task_id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "result": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "error": {
                      "type": [
                        "string",
                        "null"
                      ]
                    }
                  },
                  "required": [
                    "status"
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Authentication failure",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Task not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/gen/greet": {
      "get": {
        "summary": "Greet Endpoint",
        "description": "Greet endpoint to fetch AI model information.",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GreetResponse"
                }
              }
            }
          }
        }
      }
    },
    "/gen/init_prompt": {
      "post": {
        "summary": "Initialize Prompt Cache",
        "description": "Configure and initialize system prompt cache in Vertex AI.",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InitPromptRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultResponse"
                }
              }
            }
          }
        }
      }
    },
    "/gen/delete_cache": {
      "post": {
        "summary": "Delete Cache",
        "description": "Clean up/delete initialized system prompt cache.",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteCacheRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultResponse"
                }
              }
            }
          }
        }
      }
    },
    "/gen/check/user": {
      "post": {
        "summary": "Check User (Async)",
        "description": "Asynchronously run quality check with user parameters.",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          }
        }
      }
    },
    "/gen/check/user/sync": {
      "post": {
        "summary": "Check User (Sync)",
        "description": "Synchronously run quality check with user parameters.",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultResponse"
                }
              }
            }
          }
        }
      }
    },
    "/gen/trans/user": {
      "post": {
        "summary": "Translate User (Async)",
        "description": "Asynchronously run translation with user parameters.",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          }
        }
      }
    },
    "/gen/trans/user/sync": {
      "post": {
        "summary": "Translate User (Sync)",
        "description": "Synchronously run translation with user parameters.",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultResponse"
                }
              }
            }
          }
        }
      }
    },
    "/gen/check/default": {
      "post": {
        "summary": "Process check/default (Async)",
        "description": "Asynchronously run task for category check and prompt default.",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/gen/check/default/sync": {
      "post": {
        "summary": "Process check/default (Sync)",
        "description": "Synchronously run task for category check and prompt default.",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/gen/trans/default": {
      "post": {
        "summary": "Process trans/default (Async)",
        "description": "Asynchronously run task for category trans and prompt default.",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/gen/trans/default/sync": {
      "post": {
        "summary": "Process trans/default (Sync)",
        "description": "Synchronously run task for category trans and prompt default.",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    }
  },
  "webhooks": {}
}